+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - #2723: added read keyboardhook, gatewayaccesstoken and gatewaycredentialssource from RDP File
 - #2690: added தமிழ் (ta) Translation update
 - #2643: added registry Settings: enhancements and new settings implementation
+- #2642: added dark native title bar and matching dark mode for scrollbars, floating windows and dialogs
 - #2591: add Clickstudios Passwordstate API connector
 - #2212: added 1Password integration